requirements/test.txt
-r base.txt
coverage==3.6
django-discover-runner==0.2.2
django-dynamic-fixture==1.6.5
